src/requirements.txt
pyhcl==0.4.4
PyYAML==5.4.1